Refactor server into api/ folder to simplifying adding Authentication
