+++ api/index.js
@@ -1,0 +1,4 @@
+const {Entry} = require('./model');
+const {router} = require('./router');
+
+module.exports = {Entry, router};
